+++ sdk/monitor/azure-monitor-opentelemetry-exporter/CHANGELOG.md
@@ -14,6 +14,8 @@
     ([#33432](https://github.com/Azure/azure-sdk-for-python/pull/33432))
 - Shutdown statsbeat on customer getting 400 error code
     ([#33489](https://github.com/Azure/azure-sdk-for-python/pull/33489))
+- Track custom events extension in feature statsbeat
+    ([#33489](https://github.com/Azure/azure-sdk-for-python/pull/33489))
 
 ## 1.0.0b19 (2023-11-20)